Add configuration to opt out of removal of default reviewers
It requires extra permissions that can't be satisfied by the default `GITHUB_TOKEN`.
By making it configurable, teams can choose to use the default token and sacrifice this feature, or use a GitHub App if they want it.

+++ backport/backport.js
@@ -4,7 +4,8 @@ var __importDefault = (this && this.__importDefault) || function (mod) {
     return (mod && mod.__esModule) ? mod : { "default": mod };
 };
 Object.defineProperty(exports, "__esModule", { value: true });
-exports.backport = exports.getFinalLabels = exports.getFailedBackportCommentBody = exports.isBettererConflict = exports.LABEL_NO_CHANGELOG = exports.LABEL_ADD_TO_CHANGELOG = exports.BETTERER_RESULTS_PATH = void 0;
+exports.backport = exports.getFailedBackportCommentBody = exports.isBettererConflict = exports.LABEL_NO_CHANGELOG = exports.LABEL_ADD_TO_CHANGELOG = exports.BETTERER_RESULTS_PATH = void 0;
+exports.getFinalLabels = getFinalLabels;
 const core_1 = require("@actions/core");
 const exec_1 = require("@actions/exec");
 const github_1 = require("@actions/github");
@@ -53,7 +54,7 @@ const isBettererConflict = async (gitUnmergedPaths) => {
     return gitUnmergedPaths.length === 1 && gitUnmergedPaths[0] === exports.BETTERER_RESULTS_PATH;
 };
 exports.isBettererConflict = isBettererConflict;
-const backportOnce = async ({ base, body, commitToBackport, github, head, labelsToAdd, owner, repo, title, mergedBy, }) => {
+const backportOnce = async ({ base, body, commitToBackport, github, head, labelsToAdd, removeDefaultReviewers, owner, repo, title, mergedBy, }) => {
     const git = async (...args) => {
         await (0, exec_1.exec)('git', args, { cwd: repo });
     };
@@ -126,7 +127,7 @@ const backportOnce = async ({ base, body, commitToBackport, github, head, labels
         });
     }
     // Remove default reviewers
-    if (createRsp.data.requested_reviewers) {
+    if (removeDefaultReviewers && createRsp.data.requested_reviewers) {
         const reviewers = createRsp.data.requested_reviewers.map((user) => user.login);
         await github.pulls.deleteReviewRequest({
             pull_number: pullRequestNumber,
@@ -206,7 +207,7 @@ const getFailedBackportCommentBody = ({ base, commitToBackport, errorMessage, he
     return lines.join('\n');
 };
 exports.getFailedBackportCommentBody = getFailedBackportCommentBody;
-const backport = async ({ issue, labelsToAdd, payload: { action, label, pull_request: { labels, merge_commit_sha: mergeCommitSha, merged, number: pullRequestNumber, title: originalTitle, merged_by, }, repository: { name: repo, owner: { login: owner }, }, }, titleTemplate, token, github, sender, }) => {
+const backport = async ({ issue, labelsToAdd, payload: { action, label, pull_request: { labels, merge_commit_sha: mergeCommitSha, merged, number: pullRequestNumber, title: originalTitle, merged_by, }, repository: { name: repo, owner: { login: owner }, }, }, titleTemplate, removeDefaultReviewers, token, github, sender, }) => {
     const payload = github_1.context.payload;
     console.log('payloadAction: ' + payload.action);
     if (payload.action !== 'closed') {
@@ -322,6 +323,7 @@ const backport = async ({ issue, labelsToAdd, payload: { action, label, pull_req
                     github: github,
                     head,
                     labelsToAdd: prLabels,
+                    removeDefaultReviewers,
                     owner,
                     repo,
                     title,
@@ -387,5 +389,4 @@ function getFinalLabels(originalLabels, labelsToAdd) {
     }
     return result;
 }
-exports.getFinalLabels = getFinalLabels;
 //# sourceMappingURL=backport.js.map
